Updated FizzBuzz Program
Removed math library since it was not being used in this program. Added try and except. Exception was added to display an error message if the program was to run into an error. Changed the range in the for loop to 101 after I noticed that the program was stopping at 99 and not 100.